Update arche_constants.rdf
Added hasCustomCitation {Wallnig, Thomas and Rabl, Irene and Bisayew, Abubakar Sidyk} for the TopCollection because our cite block only shows the creators if there are some next to the PI. Wallnig would otherwise not be displayed anywhere. (The order was taken from the DA, the contributors were omitted).
Added details to orgs vemg & smelk.